Fixed PR review problems
